First step on distributed servers
Added a new message type in Server.main that processes requests for
games, it in turn sends a message to the server, announcing an incoming
player, and it sends a message to the player with information to connect
to the server.

Added the function that handles the Play button in flash. It receives
the data, and traces todo.

Server.game can create new rooms if requested by Server.matchmaker.